// src/app/generated/caster-api/model/createDesignModuleCommand.ts
/*
Copyright 2021 Carnegie Mellon University. All Rights Reserved. 
 Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.
*/

/**
 * Caster API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModuleValue } from './moduleValue';


export interface CreateDesignModuleCommand { 
    /**
     * The Id of the Design to add this DesignModule
     */
    designId?: string;
    /**
     * The Id of the selected Module for this DesignModule
     */
    moduleId?: string;
    /**
     * Name of the DesignModule.
     */
    name?: string | null;
    /**
     * Version of the selected Module to use
     */
    moduleVersion?: string | null;
    /**
     * Values for each input
     */
    values?: Array<ModuleValue> | null;
}